@@ -75,18 +75,18 @@
     "@tiptap/extension-text-align": "^2.0.0-beta.29",
     "@tiptap/extension-text-style": "^2.0.0-beta.23",
     "@tiptap/extension-underline": "^2.0.0-beta.23",
-    "@tiptap/starter-kit": "^2.0.0-beta.183",
-    "@tiptap/vue-2": "^2.0.0-beta.77",
-    "@typescript-eslint/eslint-plugin": "^5.19.0",
-    "@typescript-eslint/parser": "^5.20.0",
-    "core-js": "^3.22.0",
-    "eslint": "^8.13.0",
+    "@tiptap/starter-kit": "^2.0.0-beta.185",
+    "@tiptap/vue-2": "^2.0.0-beta.79",
+    "@typescript-eslint/eslint-plugin": "^5.25.0",
+    "@typescript-eslint/parser": "^5.25.0",
+    "core-js": "^3.22.5",
+    "eslint": "^8.15.0",
     "eslint-import-resolver-typescript": "^2.7.1",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-nuxt": "^3.2.0",
     "pug": "^3.0.2",
     "pug-plain-loader": "^1.1.0",
-    "typescript": "^4.6.3",
+    "typescript": "^4.6.4",
     "vue-class-component": "^7.2.6",
     "vue-property-decorator": "^9.1.2"
   }
