@@ -2,7 +2,7 @@
   mutation-modal-form(
     @close="close"
     :mutation="require('~/gql/dcis/mutations/document/add_document.graphql')"
-    :variables="{ comment, periodId, statusId: status && Number(status.id) }"
+    :variables="variables"
     :header="String($t('dcis.documents.add.header'))"
     :button-text="String($t('add'))"
     :update="update"
@@ -14,16 +14,27 @@
     template(#form)
       validation-provider(:name="String($t('dcis.documents.add.comment'))" rules="required" v-slot="{ errors, valid }")
         v-text-field(v-model="comment" :error-messages="errors" :success="valid" :label="$t('dcis.documents.add.comment')" autofocus)
-      validation-provider(name="String($t('dcis.documents.add.status'))" rules="required" v-slot="{ errors, valid }")
+      validation-provider(:name="String($t('dcis.documents.add.status'))" rules="required" v-slot="{ errors, valid }")
         v-combobox(v-model="status" :items="statuses" :label="$t('dcis.documents.add.status')" item-text="name" item-value="id")
+      v-combobox(v-if="documents.length" v-model="document" :items="documents" :label="$t('dcis.documents.add.lastDocument')" clearable)
+        template(#selection="{ item }") Версия {{ item.version }}
+        template(#item="{ item }")
+          v-list-item-content
+            v-list-item-title Версия {{ item.version }}
+            v-list-item-subtitle {{ item.comment }}
 </template>
 
 <script lang="ts">
 import { DataProxy } from 'apollo-cache'
 import type { PropType } from '#app'
-import { defineComponent, ref, Ref } from '#app'
 import { useCommonQuery } from '~/composables'
-import { AddDocumentMutationPayload, StatusesQuery, StatusesQueryVariables, StatusType } from '~/types/graphql'
+import {
+  AddDocumentMutationPayload, AddDocumentMutationVariables,
+  PeriodType,
+  StatusesQuery,
+  StatusesQueryVariables,
+  StatusType
+} from '~/types/graphql'
 import statusesQuery from '~/gql/dcis/queries/statuses.graphql'
 import MutationModalForm from '~/components/common/forms/MutationModalForm.vue'
 
@@ -33,11 +44,13 @@ export default defineComponent({
   components: { MutationModalForm },
   props: {
     update: { type: Function as PropType<(cache: DataProxy, result: AddDocumentMutationResultType) => void>, required: true },
-    periodId: { type: String, required: true }
+    period: { type: Object as PropType<PeriodType>, required: true },
+    documents: { type: Array as PropType<DocumentType>, default: () => ([]) }
   },
-  setup () {
-    const comment: Ref<string> = ref<string>('')
-    const status: Ref<StatusType | null> = ref<StatusType | null>(null)
+  setup (props) {
+    const comment = ref<string>('')
+    const status = ref<StatusType | null>(null)
+    const document = ref<DocumentType>(null)
 
     const { data: statuses, onResult } = useCommonQuery<StatusesQuery, StatusesQueryVariables>({
       document: statusesQuery
@@ -46,12 +59,19 @@ export default defineComponent({
       status.value = statuses[0]
     })
 
+    const variables = computed<AddDocumentMutationVariables>(() => ({
+      comment: comment.value,
+      periodId: props.period.id,
+      statusId: status.value?.id,
+      documentId: document.value?.id
+    }))
+
     const close = () => {
       comment.value = ''
       status.value = statuses[0]
     }
 
-    return { comment, status, statuses, close }
+    return { comment, status, document, statuses, variables, close }
   }
 })
 </script>
