@@ -14,7 +14,7 @@ import { useVModel } from '@vueuse/core'
 import type { Ref } from '#app'
 import { defineComponent } from '#app'
 import { useQueryRelay } from '~/composables'
-import { usePageStore } from '~/store'
+import { usePageStore } from '~/stores'
 import { CategoriesQuery, CategoriesQueryVariables, CategoryType } from '~/types/graphql'
 import categoriesQuery from '~/gql/pages/queries/categories.graphql'
 import Notification from '~/components/global/Notification.vue'
