@@ -4,11 +4,11 @@
  */
 
 import { Context, Middleware } from '@nuxt/types'
-import { useAuthStore } from '~/store/auth-store'
+import { useAuthStore } from '~/stores'
 import { MeQuery, MeQueryVariables, UserType } from '~/types/graphql'
 import meQuery from '~/gql/core/queries/me.graphql'
 
-export default <Middleware> async function ({ app: { $apolloHelpers, apolloProvider }, store }: Context) {
+export default <Middleware> async function ({ app: { $apolloHelpers, apolloProvider } }: Context) {
   const hasToken = Boolean($apolloHelpers.getToken())
   const authStore = useAuthStore()
   if (hasToken && !authStore.loginIn) {
@@ -19,10 +19,8 @@ export default <Middleware> async function ({ app: { $apolloHelpers, apolloProvi
     }).then(({ data }) => data.me as UserType)
     if (user) {
       authStore.user = user as UserType
-      await store.dispatch('auth/fetchExistUser', Object.assign({}, user)) // Убрать после удаления vuex
     } else {
       authStore.user = null
-      await store.dispatch('auth/logout') // Убрать после удаления vuex
     }
   }
 }
