@@ -1,6 +1,6 @@
 /**
- * Прослойка авторзации
- * Если пользователь не авторизован, пересылаем его на клавную
+ * Прослойка авторизации
+ * Если пользователь не авторизован, пересылаем его на главную
  */
 
 import { Middleware, Context } from '@nuxt/types'
