@@ -7,8 +7,9 @@
 </template>
 
 <script lang="ts">
-import { useNuxt2Meta } from '#app'
-import { useAuthStore } from '~/store/auth-store'
+import { defineComponent, useNuxt2Meta, useRouter } from '#app'
+import { useApolloHelpers, useI18n } from '~/composables'
+import { useAuthStore } from '~/stores'
 
 export default defineComponent({
   middleware: 'auth',
@@ -16,15 +17,12 @@ export default defineComponent({
     const router = useRouter()
     const { t, localePath } = useI18n()
     const { onLogout, defaultClient } = useApolloHelpers()
-    const { $store } = useNuxtApp()
     const userStore = useAuthStore()
 
     useNuxt2Meta({ title: t('auth.logout.logout') as string })
 
     if (userStore.loginIn) {
       userStore.logout()
-      // Убрать после удаления vuex
-      $store.dispatch('auth/logout')
       onLogout(defaultClient, true)
     }
     // Необходимо для нормальной перезагрузки сокетов
