@@ -21,7 +21,7 @@ import { DataTableHeader } from 'vuetify'
 import type { Ref, PropType } from '#app'
 import { defineComponent, inject, onMounted, ref, useRoute, useRouter, toRef } from '#app'
 import { useApolloHelpers, useFilters, useI18n } from '~/composables'
-import { HasPermissionFnType, useAuthStore } from '~/store'
+import { HasPermissionFnType, useAuthStore } from '~/stores'
 import { PeriodType, ProjectType } from '~/types/graphql'
 import { toGlobalId } from '~/services/graphql-relay'
 import { BreadCrumbsItem } from '~/types/devind'
