@@ -23,7 +23,7 @@ import type { PropType, Ref } from '#app'
 import { defineComponent, onMounted, ref, toRef, useNuxt2Meta, useRoute, useRouter } from '#app'
 import { BreadCrumbsItem } from '~/types/devind'
 import { useApolloHelpers, useCursorPagination, useFilters, useI18n, useQueryRelay } from '~/composables'
-import { HasPermissionFnType, useAuthStore } from '~/store'
+import { HasPermissionFnType, useAuthStore } from '~/stores'
 import projectsQuery from '~/gql/dcis/queries/projects.graphql'
 import { ProjectsQuery, ProjectsQueryVariables, ProjectType } from '~/types/graphql'
 import BreadCrumbs from '~/components/common/BreadCrumbs.vue'
