@@ -19,7 +19,7 @@ import { ApolloQueryResult } from '@apollo/client'
 import type { ComputedRef } from '#app'
 import { Ref, computed, defineComponent, onUnmounted, toRef, useNuxt2Meta, useRoute } from '#app'
 import { useCommonQuery, useI18n, usePage } from '~/composables'
-import { HasPermissionFnType, useAuthStore, usePageStore } from '~/store'
+import { HasPermissionFnType, useAuthStore, usePageStore } from '~/stores'
 import { PageQuery, PageQueryVariables } from '~/types/graphql'
 import { BreadCrumbsItem } from '~/types/devind'
 import pageQuery from '~/gql/pages/queries/page.graphql'
