@@ -7,7 +7,7 @@ import { ApolloQueryResult } from '@apollo/client'
 import type { ComputedRef } from '#app'
 import { computed, defineComponent, onUnmounted, useRoute } from '#app'
 import { useCommonQuery, useI18n, usePage } from '~/composables'
-import { usePageStore } from '~/store'
+import { usePageStore } from '~/stores'
 import { PageQuery, PageQueryVariables } from '~/types/graphql'
 import { BreadCrumbsItem } from '~/types/devind'
 import pageQuery from '~/gql/pages/queries/page.graphql'
