@@ -16,7 +16,7 @@ import type { Ref } from '#app'
 import { defineComponent, toRef, useNuxt2Meta } from '#app'
 import type { UserType } from '~/types/graphql'
 import { useFilters, useI18n } from '~/composables'
-import { useAuthStore } from '~/store/auth-store'
+import { useAuthStore } from '~/stores'
 import AvatarView from '~/components/users/AvatarView.vue'
 import BasicInformation from '~/components/profile/BasicInformation.vue'
 import AdditionalInformation from '~/components/profile/AdditionalInformation.vue'
