@@ -64,7 +64,7 @@
 import { DataTableHeader } from 'vuetify'
 import { useMutation } from '@vue/apollo-composable'
 import { useNuxt2Meta } from '#app'
-import { useAuthStore } from '~/store'
+import { useAuthStore } from '~/stores'
 import { ChangePasswordMutationPayload, SessionsQuery, SessionsQueryVariables, SessionType } from '~/types/graphql'
 import sessionQuery from '~/gql/core/queries/sessions.graphql'
 import deleteSessions from '~/gql/core/mutations/user/delete_sessions.graphql'
