@@ -1,7 +1,7 @@
 import graphene
 
 from apps.dcis.schema.queries import DocumentQueries, PrivilegeQueries, ProjectQueries, SheetQueries
-from apps.dcis.schema.mutations import DocumentMutations, ProjectMutations, SheetMutations, PrivilegeMutations
+from apps.dcis.schema.mutations import CellMutations, DocumentMutations, ProjectMutations, SheetMutations, PrivilegeMutations
 
 
 class Query(
@@ -17,6 +17,7 @@ class Query(
 
 
 class Mutation(
+    CellMutations,
     ProjectMutations,
     DocumentMutations,
     SheetMutations,
