@@ -1,18 +1,27 @@
-from typing import List
+from typing import Any
+
 import graphene
+from devind_helpers.orm_utils import get_object_or_404
 from django.db.models import QuerySet
 from graphene_django import DjangoListField
+from graphene_django.filter import DjangoFilterConnectionField
 from graphql import ResolveInfo
 from graphql_relay import from_global_id
-from devind_helpers.orm_utils import get_object_or_404
 
 from apps.dcis.models import Document, DocumentStatus
 from apps.dcis.schema.types import DocumentType, StatusType, DocumentStatusType
+from apps.dcis.services.document_services import get_documents
 
 
 class DocumentQueries(graphene.ObjectType):
     """Запросы записей, связанных с документами."""
-
+    documents = DjangoFilterConnectionField(
+        DocumentType,
+        period_id=graphene.ID(required=True, description='Идентификатор периода'),
+        divisions_id=graphene.List(graphene.NonNull(graphene.Int), description='Идентификаторы дивизионов'),
+        required=True,
+        description='Документы'
+    )
     document = graphene.Field(
         DocumentType,
         description='Документ',
@@ -25,6 +34,11 @@ class DocumentQueries(graphene.ObjectType):
         document_id=graphene.ID(description='Идентификатор документа'),
         description='Статусы документов'
     )
+
+    @staticmethod
+    def resolve_documents(root: Any, info: ResolveInfo, period_id: str, divisions_id: list[str] = []):
+        period_id: int = from_global_id(period_id)[1]
+        return get_documents(info.context.user, period_id, divisions_id)
 
     @staticmethod
     def resolve_document(root, info: ResolveInfo, document_id: str, *args, **kwargs) -> Document:
