@@ -1,4 +1,4 @@
-from typing import Any
+from typing import Any, Optional
 
 import graphene
 from devind_helpers.decorators import permission_classes
@@ -9,8 +9,9 @@ from graphene_django_filter import AdvancedDjangoFilterConnectionField
 from graphql import ResolveInfo
 from graphql_relay import from_global_id
 
+from apps.core.models import User
 from apps.dcis.models import Project, Period
-from ..types import ProjectType, PeriodType
+from ..types import DivisionType, DivisionModelType, ProjectType, PeriodType
 
 
 class ProjectQueries(graphene.ObjectType):
@@ -36,6 +37,20 @@ class ProjectQueries(graphene.ObjectType):
         description='Информация по периоду'
     )
 
+    period_divisions = AdvancedDjangoFilterConnectionField(DivisionType, description='Получение дивизионов')
+    user_divisions = graphene.List(
+        DivisionModelType,
+        user_id=graphene.ID(description='Пользователь'),
+        project_id=graphene.ID(description='Идентификатор проекта'),
+        required=True,
+        description='Дивизионы пользователя'
+    )
+
+    @staticmethod
+    @permission_classes((IsAuthenticated,))
+    def resolve_periods(root: Any, info: ResolveInfo, user_id: str, *args, **kwargs):
+        return set(Period.objects.filter(periodprivilege__user=from_global_id(user_id)[1]).all())
+
     @staticmethod
     @permission_classes((IsAuthenticated,))
     def resolve_project(root: Any, info: ResolveInfo, project_id: str, *args, **kwargs):
@@ -48,5 +63,17 @@ class ProjectQueries(graphene.ObjectType):
 
     @staticmethod
     @permission_classes((IsAuthenticated,))
-    def resolve_periods(root: Any, info: ResolveInfo, user_id: str, *args, **kwargs):
-        return set(Period.objects.filter(periodprivilege__user=from_global_id(user_id)[1]).all())
+    def resolve_user_divisions(
+        root: Any,
+        info: ResolveInfo,
+        user_id: Optional[str] = None,
+        project_id: Optional[str] = None
+    ) -> list:
+        user: User = info.context.user \
+            if user_id is None \
+            else get_object_or_404(User, pk=from_global_id(user_id)[1])
+        project: Optional[Project] = None \
+            if project_id is None \
+            else get_object_or_404(Project, pk=from_global_id(project_id)[1])
+        divisions = user.divisions(project)
+        return divisions
